code synchronization at 2018-01-05@19:16

+++ framework/fota/platform/ota_platform_os.h
@@ -20,7 +20,7 @@ void ota_mutex_destroy(void *mutex);
 
 void *ota_semaphore_init(void);
 
-int8_t ota_semaphore_wait(void *sem, uint32_t timeout_ms);
+int ota_semaphore_wait(void *sem, int32_t timeout_ms);
 
 void ota_semaphore_post(void *sem);
